Fixed lock-unlock scheme in readByte/readBit (they call each other, unlocking prior calling the other one). Fixed an off-by-one write error in readBits.

+++ modules/native/feathers/bitbuf/bitbuf_ext.cpp
@@ -280,7 +280,7 @@ static void set_wpos( const Class*, const String&, void* instance, const Item& v
       throw FALCON_SIGN_ERROR( RangeError, e_param_type );
    }
 
-   buf->wpos( pos );
+   buf->wpos( (uint32) pos );
 }
